Add formatting tests to travis
We currently check code formatting and remind people to obey style by hand.
This is tedious and error-prone, made worse by github’s unusual rendering,
and badly-formatted code has made it into our repository. This commit adds
a travis job which checks PRs for correct formatting, so we no
longer need humans to check for this in the common case.